chore: move logger to separate module
