+++ app/common/components/Session/Session.jsx
@@ -8,7 +8,7 @@ import {BUTTON} from '../../constants/antd-types';
 import {LINKS} from '../../constants/common';
 import {ADD_CLOUD_PROVIDER_TAB_KEY} from '../../constants/session-builder';
 import {ipcRenderer, shell} from '../../shared/polyfills';
-import {log} from '../../utils/other';
+import {log} from '../../utils/logger';
 import AdvancedServerParams from './AdvancedServerParams.jsx';
 import AttachToSession from './AttachToSession.jsx';
 import CapabilityEditor from './CapabilityEditor.jsx';
